VS build: cleanup settings and actually include the new files.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@16538 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/dbc/dbc.hpp
@@ -445,7 +445,7 @@ public:
   { return _id != 0; }
 
   // Composite functions
-  const spelleffect_data_t& effectN( unsigned idx ) const
+  const spelleffect_data_t& effectN( size_t idx ) const
   {
     assert( _effects );
     assert( idx > 0 && "effect index must not be zero or less" );
